5022.104 - SDK update

+++ MADNET_SDK_DEMO/CTableADViewController.m
@@ -108,7 +108,7 @@
     {
         static NSString *kCellIdentifier = @"Cell";
         cell = [tableView dequeueReusableCellWithIdentifier:kCellIdentifier forIndexPath:indexPath];
-        [cell.textLabel setText:[NSString stringWithFormat: @"Simple Test Cell %d", indexPath.row]];
+        [cell.textLabel setText:[NSString stringWithFormat: @"Simple Test Cell %ld", (long)indexPath.row]];
     }
 
     return cell;
